pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Difference generator is a program which defines the difference between two data structures."
authors = ["Anastasiia Nezgovorova <odabcrfake@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

